Nested the model file to give me more space
