+++ mage_game/model/__init__.py
@@ -1,0 +1,3 @@
+from .magic import *
+from .character import *
+from .state import *
